b4pull

+++ FintrakBanking.ViewModels/Credit/LoanViewModel.cs
@@ -297,7 +297,7 @@ namespace FintrakBanking.ViewModels.Credit
         public decimal interesrtOnPastDueInterest { get; set; }
         public decimal penalChargeAmount { get; set; }
         public DateTime? lastRestructureDate  { get; set; }
-        public short loanSystemTypeId  { get; set; }
+        public int loanSystemTypeId  { get; set; }
         public bool isPerforming { get; set; }
         public object loanPrincipal { get; set; }
         public List<LoanDisbursementViewModel> loanBeneficiary { get; set; }
@@ -1483,7 +1483,7 @@ namespace FintrakBanking.ViewModels.Credit
         public string loanRefNo { get; set; }
         public string customerName { get; set; }
         public int loanId { get; set; }
-        public short loanSystemTypeId { get; set; }
+        public int loanSystemTypeId { get; set; }
     }
 
 
@@ -1571,7 +1571,7 @@ namespace FintrakBanking.ViewModels.Credit
         
         public int pastDueDays { get; set; }
         public int loanId { get; set; }
-        public short loanSytemTypeId { get; set; }
+        public int loanSytemTypeId { get; set; }
         public string businessDevelopmentManger { get; set; }
     }
